fix: Remove all wallet
